from request macro module
